Fix duplicated classses in sample addon doc
Clear the LOCAL_DROIDDOC_OPTIONS which was originally assigned with
package name. This variable is assumed to be set with local javadoc
options. 

When it is set with the package name, Javadoc will take both the 
packages in this variable and droiddoc-src-list generated by 
build/core/droiddoc.mk as its individual input at the same time and 
build the html-doc with duplicate classes.

You can get some details of this problem in Issue 74070

Signed-off-by: Wenda song <wenda.song@intel.com>
Change-Id: I5aa5a4de92a77cf6ff5e6c29b60e4931623d2298